backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
email-validator
sqlalchemy
psycopg2-binary
python-multipart
python-jose[cryptography]
passlib[bcrypt]
langchain
langchain-groq
langchain-huggingface
langchain-community
langchain-pinecone
sentence-transformers
pinecone-client
qdrant-client
langchain-qdrant
pypdf
docx2txt
pandas
python-dotenv
fpdf2
